김관호 : 충돌체 생성 및 GDI 클래스 생성

+++ WinAPI/Client/CPathMgr.cpp
@@ -16,7 +16,7 @@ void CPathMgr::Init()
 {
 	GetCurrentDirectory(255, m_szContentPath);
 
-	int iLen = wcslen(m_szContentPath);
+	int iLen = static_cast<int>(wcslen(m_szContentPath));
 
 	// 상위 폴더로
 	for (int i = iLen - 1; i >= 0; i--)
